jkpRegsitrationPOCgrpc/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "JKP Satsangi Registration POC - gRPC Backend"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "protobuf>=5.29.0",
    "pydantic==2.12.5",
    "ruff==0.15.5",
    "fastapi==0.135.1",
    "uvicorn==0.41.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
